+++ lib/src/layouts/languages/languages.dart
@@ -12,3 +12,4 @@ library;
 
 export 'bengali.dart';
 export 'english.dart';
+export 'french.dart';